+++ app/src/main/java/org/simple/clinic/overdue/download/OverdueListDownloadResult.kt
@@ -1,0 +1,12 @@
+package org.simple.clinic.overdue.download
+
+import android.net.Uri
+
+sealed class OverdueListDownloadResult {
+
+  data class DownloadSuccessful(val uri: Uri) : OverdueListDownloadResult()
+
+  object NotEnoughStorage : OverdueListDownloadResult()
+
+  object DownloadFailed : OverdueListDownloadResult()
+}