+++ app/src/main/java/org/simple/clinic/overdue/download/formatdialog/SelectOverdueDownloadFormatViewEffectHandler.kt
@@ -11,6 +11,8 @@ class SelectOverdueDownloadFormatViewEffectHandler(
     when (viewEffect) {
       is ShareDownloadedFile -> uiActions.shareDownloadedFile(viewEffect.uri, viewEffect.mimeType)
       Dismiss -> uiActions.dismiss()
+      OpenNotEnoughStorageErrorDialog -> uiActions.openNotEnoughStorageErrorDialog()
+      OpenDownloadFailedErrorDialog -> uiActions.openDownloadFailedErrorDialog()
     }.exhaustive()
   }
 }