+++ app/src/test/java/org/simple/clinic/overdue/download/formatdialog/SelectOverdueDownloadFormatUpdateTest.kt
@@ -9,6 +9,9 @@ import com.spotify.mobius.test.NextMatchers.hasNoModel
 import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
+import org.simple.clinic.overdue.download.OverdueListDownloadResult.DownloadFailed
+import org.simple.clinic.overdue.download.OverdueListDownloadResult.DownloadSuccessful
+import org.simple.clinic.overdue.download.OverdueListDownloadResult.NotEnoughStorage
 import org.simple.clinic.overdue.download.OverdueListFileFormat.CSV
 import org.simple.clinic.overdue.download.OverdueListFileFormat.PDF
 
@@ -51,7 +54,7 @@ class SelectOverdueDownloadFormatUpdateTest {
 
     updateSpec
         .given(defaultModel)
-        .whenEvent(FileDownloadedForSharing(uri))
+        .whenEvent(FileDownloadedForSharing(DownloadSuccessful(uri)))
         .then(assertThatNext(
             hasModel(defaultModel.overdueDownloadCompleted()),
             hasEffects(ShareDownloadedFile(uri, defaultModel.overdueListFileFormat.mimeType))
@@ -101,6 +104,28 @@ class SelectOverdueDownloadFormatUpdateTest {
         .then(assertThatNext(
             hasNoModel(),
             hasNoEffects()
+        ))
+  }
+
+   @Test
+  fun `when there is not enough space to download the file, then open not enough storage error dialog`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(FileDownloadedForSharing(NotEnoughStorage))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(OpenNotEnoughStorageErrorDialog)
+        ))
+  }
+
+  @Test
+  fun `when download fails, then open download failed error dialog`() {
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(FileDownloadedForSharing(DownloadFailed))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(OpenDownloadFailedErrorDialog)
         ))
   }
 }