+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   - Enable incremental annotation processing
 - Use `IO` scheduler for doing background tasks in `OverdueDownloadWorker`
 - Download/Share CSV only for API below 24
+- Show error dialog when overdue list download fails
 
 ### Changes